move exported code to pkg
